Reintroduce scanner.* (removed in commit ba0835c)
This seems to do some metadata related things to the .3dsx that may be
important in some cases (like the bug described by Xenoisaga in
GBAtemp, hope it fixes it). Needs to change some things in boot.* to allow
.3dsx file scan.

Resync changes with Homebrew Launcher too, so any bug fixes that
@smealum may did in HBL will be fixed here too.

Anyway, the advantage of a clean source code is that now I somewhat
understand what scanner.* files does, so it made easier to reintroduce
it without including things like tinyxml2.