Reactivate warnings, increase default delay
I fixed all warnings from libcakebrah, so I can remove the -w CFLAG and
readd the -Wall. Thanks to this, identified one missing include in
misc.h.

With payload loader it seems that 100 ms is not sufficient to have
realible boot, however with 150 ms it seems ok, at least in my 3DS.