Fix payload loading, payload should default to false
It seems that the order of function calls really matters in this case.
svcSleepThread should be the first function to be called. I thought that
calling it right after firm_reboot() or brahma_exit() would be better,
however it should be the first function to be called. And you need to
call exit_services() before calling firm_reboot(): this was pretty
obvious, however I was tired and didn't see it.

The worst part is that if I looked at the source code from
Cpasjuste/CtrBootManager it would be very easily to realize how was the
right way to make it work. Anyway, it is working now.

The second problem this commit fixes is actually a partial revert of
commit 17a95eb. Since the [DEFAULT] section would overwrite payload
variable too, it would by consequence trying to load everything as a
payload if payload in [DEFAULT] was true.

There was two ways to fixes this, the easier one was to just drop the
behavior from commit 17a95eb. The other one would maintain the behavior,
however would be ugly and whatever, this works.