Payload loading using CakeBrah finally works
Who would knows that you actually need to initialize the GFX to make
CakeBrah/libkhax works? For now it only seems to work reliable in
CakesFW, rxTools seems to have a really low boot rate (or none at all).

Probably it should be something simple, maybe looking for the official
repository of rxTools to understand what they did with CakeBrah.