Using upstream CakeBrah instead of libcakebrah fork
Upstream fixed lots of reported issues, so I think it is time to change
to them. I still think that they organization of repository makes a mess
in Makefile, however it is easily fixed anyway (just not pretty).

Don't forget to do a "make clean" before building after pulling this
changes, or build may fail!