Voodoo to improve rxTools boot rate
I don't know wtf why rxTools is the only CFW that does not boot reliable
with the old delay method. So I created a Frankstein with the old delay
method and the one used by CtrBootManager. Hope this brings the best of
both worlds (reliable boot rate in rxTools and delay configurable in
ms).